fix: commentLikeCount 타입 Integer → Long 변경

+++ src/main/java/com/springboot/monew/comment/dto/CommentLikeDto.java
@@ -15,5 +15,5 @@ public record CommentLikeDto(
     UUID commentUserId,
     String commentUserNickname,
     String commentContent,
-    Integer commentLikeCount,
+    Long commentLikeCount,
     Instant commentCreatedAt) {}
